Fix evidence packet transmission logic

+++ CM4/Core/Src/opusenc.c
@@ -54,7 +54,7 @@ void opusenc_encode(const int16_t* restrict audio_in, const opus_frame_t** restr
       // If encoding buffer is full, encode the audio data frame
       if (!encode_buffer_index)
       {
-         opus_frame->num_encoded_bytes = (uint16_t)opus_encode(&opus_encoder, encode_buffer, opus_frame->encoded_data, sizeof(opus_frame->encoded_data));
+         opus_frame->num_encoded_bytes = (uint8_t)opus_encode(&opus_encoder, encode_buffer, opus_frame->encoded_data, sizeof(opus_frame->encoded_data));
          *result_end = opus_frame = opus_frame->next;
       }
    }
